requirements.txt
pytesseract
opencv-python
numpy
websockets